refactor(app): extract status projection
